requirements.txt
MouseInfo==0.1.3
numpy==1.26.2
opencv-python==4.8.1.78
Pillow==10.1.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
pytweening==1.0.7chess==1.10.0
imageio==2.33.0
lazy_loader==0.3
MouseInfo==0.1.3
networkx==3.2.1
numpy==1.26.2
opencv-python==4.8.1.78
packaging==23.2
Pillow==10.1.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python-chess==1.999
pytweening==1.0.7
scikit-image==0.22.0
scipy==1.11.4
tifffile==2023.9.26
